requirements.txt
apscheduler
discord.py
pyyaml
termcolor
